requirements.txt
youtube-comment-downloader
beautifulsoup4
fastmcp
yt_dlp